feat: Commit 1.7-DEV: Fix SecretCode architecture engine, command, listeners, controllers upd..

+++ src/main/java/tvgirl/elmodev/e1m0Admin/api/repo/gui/SecretCodeRepositoryAPI.java
@@ -9,6 +9,7 @@ public interface SecretCodeRepositoryAPI {
 
     /* SYSTEM | 💾 */
     void systemSetSecretCode(UUID adminID, int code);
-    byte getSecretCode(UUID adminID);
+
+    int getSecretCode(UUID adminID);
 
 }
